blog/requirements.txt
Django
djangorestframework
drf_yasg
djangorestframework_simplejwt
django_filter